+++ cg_lims/EPPs/files/base.py
@@ -3,8 +3,7 @@ import click
 
 # commands
 from cg_lims.EPPs.files.file_to_udf import csv_well_to_udf
-from cg_lims.EPPs.files.hamilton_normalization_file import make_hamilton_barcode_file
-from cg_lims.EPPs.files.make_kapa_csv import make_kapa_csv
+from cg_lims.EPPs.files.hamilton.base import hamilton
 from cg_lims.EPPs.files.pooling_map.make_pooling_map import pool_map
 from cg_lims.EPPs.files.placement_map.make_96well_placement_map import placement_map
 
@@ -16,8 +15,7 @@ def files(ctx):
     pass
 
 
-files.add_command(make_kapa_csv)
 files.add_command(csv_well_to_udf)
 files.add_command(pool_map)
 files.add_command(placement_map)
-files.add_command(make_hamilton_barcode_file)
+files.add_command(hamilton)
