+++ cg_lims/get/fields.py
@@ -1,7 +1,9 @@
 import datetime as dt
+
+from genologics.descriptors import LocationDescriptor
 from requests.exceptions import HTTPError
-from genologics.entities import Sample
-from typing import Optional
+from genologics.entities import Sample, Artifact
+from typing import Optional, Tuple
 import logging
 
 LOG = logging.getLogger(__name__)
@@ -55,3 +57,30 @@ def get_processing_time(sample: Sample) -> Optional[dt.datetime]:
         % sample.id
     )
     return None
+
+
+def get_artifact_well(artifact: Artifact) -> str:
+    """Parsing out the well position from LocationDescriptor"""
+
+    location: Tuple = artifact.location
+    return location[1].replace(":", "")
+
+
+def get_index_well(artifact: Artifact):
+    """Parsing out the index well position from the reagent label string wich
+    typically looks like this: 'A05 IDT_10nt_446 (AGCGTGACCT-CCATCCGAGT)'
+    """
+
+    if artifact.reagent_labels:
+        # Assuming one reagent label per artifact (reagent_labels is a list):
+        reagent_label = artifact.reagent_labels[0]
+
+        # Getting the index well:
+        index_well_with_zero = reagent_label.split(" ")[0]
+
+        # Picking out column and removing zeros by int():
+        index_well_col = int(index_well_with_zero[1:])
+        index_well_row = index_well_with_zero[0]
+        return f"{index_well_row}{index_well_col}"
+    else:
+        return "-"
