+++ tests/EPPs/test_make_kapa_csv.py
@@ -5,9 +5,8 @@ from genologics.entities import Artifact
 
 import pytest
 
-from cg_lims.EPPs.files.make_kapa_csv import get_file_data_and_write
+from cg_lims.EPPs.files.hamilton.make_kapa_csv import get_file_data_and_write
 from cg_lims.exceptions import MissingArtifactError, MissingUDFsError
-
 
 
 def test_get_file_data_and_write_KAPA_library_preparation(kapa_library_preparation_file, lims):
@@ -16,11 +15,11 @@ def test_get_file_data_and_write_KAPA_library_preparation(kapa_library_preparati
     # The two samples are now in another step where they have location and
     # reagent labels set.
     # (See entety_json_data for details)
-    server('test_make_kapa_csv')
+    server("test_make_kapa_csv")
     file_name = "some_file_name"
     amount_step = "Aliquot samples for enzymatic fragmentation TWIST v2"
     file = Path(file_name)
-    curent_step_artifacts = [Artifact(lims, id='2-1155129'), Artifact(lims, id='2-1155130')]
+    curent_step_artifacts = [Artifact(lims, id="2-1155129"), Artifact(lims, id="2-1155130")]
 
     # WHEN running get_file_data_and_write
     get_file_data_and_write(lims, amount_step, curent_step_artifacts, file)
@@ -38,13 +37,11 @@ def test_get_file_data_and_write_Enzymatic_fragmentation(enzymatic_fragmentation
     # The two samples are now in another step where they have location but no
     # reagent labels set.
     # (See entety_json_data for details)
-    server('test_make_kapa_csv')
+    server("test_make_kapa_csv")
     file_name = "some_file_id"
     amount_step = "Aliquot samples for enzymatic fragmentation TWIST v2"
     file = Path(file_name)
-    curent_step_artifacts = [Artifact(lims, id='2-1155124'), Artifact(lims, id='2-1155125')]
-
-
+    curent_step_artifacts = [Artifact(lims, id="2-1155124"), Artifact(lims, id="2-1155125")]
 
     # WHEN running get_file_data_and_write
     get_file_data_and_write(lims, amount_step, curent_step_artifacts, file)
@@ -62,11 +59,11 @@ def test_get_file_data_and_write_missing_udf(kapa_library_preparation_file_missi
     # The two samples are now in another step where they have location and
     # reagent labels set.
     # (See entety_json_data for details)
-    server('make_kapa_csv_missing_udfs')
+    server("make_kapa_csv_missing_udfs")
     file_name = "some_file_name"
     amount_step = "Aliquot samples for enzymatic fragmentation TWIST v2"
     file = Path(file_name)
-    curent_step_artifacts = [Artifact(lims, id='2-1155129'), Artifact(lims, id='2-1155130')]
+    curent_step_artifacts = [Artifact(lims, id="2-1155129"), Artifact(lims, id="2-1155130")]
 
     # WHEN running get_file_data_and_write
     # THEN MissingUDFsError is being raised, but the file is still created for the sample that
@@ -77,18 +74,20 @@ def test_get_file_data_and_write_missing_udf(kapa_library_preparation_file_missi
     file.unlink()
 
 
-def test_get_file_data_and_write_missing_artifact(kapa_library_preparation_file_missing_udf, lims
-):
+def test_get_file_data_and_write_missing_artifact(kapa_library_preparation_file_missing_udf, lims):
     # GIVEN: A file name. and a lims with two samples that has NOT been run
     # through a process of type <amount_step>.
     # The two samples are now in another step where they have location and
     # reagent labels set.
     # (See entety_json_data for details)
-    server('make_kapa_csv_missing_udfs')
+    server("make_kapa_csv_missing_udfs")
     file_name = "some_file_name"
     amount_step = "Aliquot samples for enzymatic fragmentation TWIST v2"
     file = Path(file_name)
-    curent_step_artifacts = [Artifact(lims, id='2-1155129_other_sample'), Artifact(lims, id='2-1155130_other_sample')]
+    curent_step_artifacts = [
+        Artifact(lims, id="2-1155129_other_sample"),
+        Artifact(lims, id="2-1155130_other_sample"),
+    ]
 
     # WHEN running get_file_data_and_write
     # THEN MissingArtifactError is being raised, and no file is being created.
